Modify the model_training.py to incorporate class imbalance
