+++ scripts/data_analysis_preprocessing.py
@@ -3,7 +3,6 @@ Data Analysis and Preprocessing
 """
 import numpy as np
 import pandas as pd
-import ipaddress
 
 import matplotlib.pyplot as plt
 import seaborn as sns
@@ -14,7 +13,6 @@ from sklearn.preprocessing import LabelEncoder, StandardScaler, OneHotEncoder
 from scipy.stats import chi2_contingency
 
 
-from scripts.smotified_gan_balancer import SMOTifiedGANBalancer
 
 
 class FraudDataProcessor:
@@ -26,25 +24,6 @@ class FraudDataProcessor:
         :param data: Pandas DataFrame
         """
         self.data = pd.read_csv(data_path)
-        self.balancer = SMOTifiedGANBalancer()
-
-    def balance_data(self, target_col='class'):
-        """
-        Calls the SMOTified+GAN method to balance data.
-        """
-        print("Balancing data using SMOTified+GAN...")
-
-        numerical_cols = self.retrieve_numerical_columns()
-        X = self.data[numerical_cols].values
-        y = self.data[target_col].values
-
-        X_balanced, y_balanced = self.balancer.balance_data(X, y)
-
-        # Store the balanced data
-        self.data = pd.DataFrame(X_balanced, columns=numerical_cols)
-        self.data[target_col] = y_balanced
-
-        print("Data Balancing Completed Successfully.")
 
     def overview_of_data(self):
         """Provide an overview of the dataset."""
